requirements.txt
tensorflow==2.15.0
streamlit==1.32.0
numpy==1.26.4
opencv-python==4.9.0.80
matplotlib==3.8.3
reportlab==4.2.0
pyyaml==6.0.1
pillow==10.2.0
tensorflow-datasets==4.9.4
pandas
numpy
seaborn